Bump version to 1.5, upgraded all packages version to latest
