+++ app/components/icons/index.ts
@@ -8,10 +8,10 @@ export { GitHubIcon } from './GitHubIcon';
 export { SiRust, SiPython, SiCplusplus, SiPhp } from 'react-icons/si';
 export { SiReact, SiNextdotjs, SiTailwindcss, SiExpress, SiFlask } from 'react-icons/si';
 export { SiRedis, SiMysql, SiSqlite } from 'react-icons/si';
-export { SiDocker, SiAmazon, SiGooglecloud, SiGit } from 'react-icons/si';
+export { SiDocker, SiGooglecloud, SiGit } from 'react-icons/si';
 export { SiArchlinux, SiGnubash } from 'react-icons/si';
 
-export { FaWindows } from 'react-icons/fa';
+export { FaWindows, FaAws } from 'react-icons/fa';
 // export { VscCode } from 'react-icons/vsc';
 
 export { RefreshCw } from 'lucide-react';
