+++ next.config.mjs
@@ -1,0 +1,57 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+	// ESLint configuration
+	eslint: {
+		ignoreDuringBuilds: true,
+	},
+	// Image optimization
+	images: {
+		unoptimized: false,
+		remotePatterns: [
+			{ protocol: 'https', hostname: 'avatars.githubusercontent.com' },
+			{ protocol: 'https', hostname: 'githubusercontent.com' },
+			{ protocol: 'https', hostname: 'opengraph.githubassets.com' },
+			{ protocol: 'https', hostname: 'images.weserv.nl' }
+		],
+		deviceSizes: [640, 750, 828, 1080, 1200, 1920, 2048, 3840],
+		imageSizes: [16, 32, 48, 64, 96, 128, 256, 384],
+		formats: ['image/webp'],
+		minimumCacheTTL: 60,
+	},
+
+	// Development + production options
+	reactStrictMode: true,
+	poweredByHeader: false,
+	compress: true,
+	productionBrowserSourceMaps: false,
+
+	// Security headers
+	async headers() {
+		return [
+			{
+				source: '/:path*',
+				headers: [
+					{ key: 'X-DNS-Prefetch-Control', value: 'on' },
+					{ key: 'Strict-Transport-Security', value: 'max-age=63072000; includeSubDomains; preload' },
+					{ key: 'X-XSS-Protection', value: '1; mode=block' },
+					{ key: 'X-Frame-Options', value: 'SAMEORIGIN' },
+					{ key: 'X-Content-Type-Options', value: 'nosniff' },
+					{ key: 'Referrer-Policy', value: 'strict-origin-when-cross-origin' },
+					{ key: 'Permissions-Policy', value: 'camera=(), microphone=(), geolocation=()' },
+				],
+			},
+		];
+	},
+
+	// Rewrite rule
+	async rewrites() {
+		return [
+			{
+				source: '/api/:path*',
+				destination: '/api/:path*',
+			},
+		];
+	},
+};
+
+export default nextConfig;
