+++ package.json
@@ -1,46 +1,41 @@
 {
-  "name": "portfolio",
-  "version": "1.0.6",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
-    "predev": "next dev --turbo"
-  },
-  "dependencies": {
-    "@formspree/react": "^3.0.0",
-    "@radix-ui/react-progress": "^1.1.7",
-    "@radix-ui/react-slot": "^1.2.3",
-    "@radix-ui/react-tabs": "^1.1.12",
-    "@radix-ui/react-toast": "^1.2.14",
-    "@tailwindcss/postcss": "^4.1.11",
-    "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
-    "eslint": "^9.31.0",
-    "gsap": "^3.13.0",
-    "lucide-react": "^0.525.0",
-    "next": "^15.4.2",
-    "next-themes": "^0.4.6",
-    "postcss": "^8.5.6",
-    "react": "19.1.0",
-    "react-dom": "19.1.0",
-    "react-hot-toast": "^2.5.2",
-    "react-icons": "^5.5.0",
-    "react-turnstile": "^1.1.4",
-    "sharp": "^0.34.3",
-    "sonner": "^2.0.6",
-    "tailwind-merge": "^3.3.1",
-    "tailwindcss": "^4.1.11",
-    "tailwindcss-animate": "^1.0.7",
-    "typescript": "5.8.3",
-    "zod": "^4.0.5"
-  },
-  "devDependencies": {
-    "@types/node": "24.0.15",
-    "@types/react": "19.1.8",
-    "@types/react-dom": "19.1.6",
-    "eslint-config-next": "^15.4.2"
-  }
+	"name": "portfolio",
+	"version": "1.5.0",
+	"private": true,
+	"scripts": {
+		"dev": "next dev --turbo",
+		"build": "next build",
+		"start": "next start",
+		"lint": "eslint app"
+	},
+	"dependencies": {
+		"@formspree/react": "^3.0.0",
+		"@radix-ui/react-progress": "^1.1.8",
+		"@radix-ui/react-slot": "^1.2.4",
+		"@radix-ui/react-toast": "^1.2.15",
+		"class-variance-authority": "^0.7.1",
+		"clsx": "^2.1.1",
+		"gsap": "^3.15.0",
+		"lucide-react": "^1.14.0",
+		"next": "^16.2.5",
+		"react": "^19.2.6",
+		"react-dom": "^19.2.6",
+		"react-icons": "^5.6.0",
+		"react-turnstile": "^1.1.5",
+		"sharp": "^0.34.5",
+		"tailwind-merge": "^3.5.0",
+		"tailwindcss": "^4.2.4",
+		"zod": "^4.4.3"
+	},
+	"devDependencies": {
+		"@eslint/eslintrc": "^3.3.5",
+		"@tailwindcss/postcss": "^4.2.4",
+		"@types/node": "^25.6.0",
+		"@types/react": "^19.2.14",
+		"@types/react-dom": "^19.2.3",
+		"eslint": "9",
+		"eslint-config-next": "^16.2.5",
+		"postcss": "^8.5.14",
+		"typescript": "^6.0.3"
+	}
 }
