--- pnpm-workspace.yaml
@@ -1,4 +1,0 @@
-ignoredBuiltDependencies:
-  - '@tailwindcss/oxide'
-  - sharp
-  - unrs-resolver
